Move communication channel constants to interaction app
Also stops using the constants in the interaction factories in tests.
