+++ datahub/interaction/test/factories.py
@@ -5,7 +5,7 @@ from datahub.company.test.factories import AdviserFactory, CompanyFactory, Conta
 from datahub.core import constants
 from datahub.core.test_utils import random_obj_for_model
 from datahub.event.test.factories import EventFactory
-from datahub.interaction.models import Interaction, ServiceDeliveryStatus
+from datahub.interaction.models import CommunicationChannel, Interaction, ServiceDeliveryStatus
 from datahub.investment.test.factories import InvestmentProjectFactory
 
 
@@ -32,7 +32,9 @@ class CompanyInteractionFactory(InteractionFactoryBase):
 
     kind = Interaction.KINDS.interaction
     company = factory.SubFactory(CompanyFactory)
-    communication_channel_id = constants.CommunicationChannel.face_to_face.value.id
+    communication_channel = factory.LazyFunction(
+        lambda: random_obj_for_model(CommunicationChannel)
+    )
 
 
 class InvestmentProjectInteractionFactory(InteractionFactoryBase):
@@ -40,7 +42,9 @@ class InvestmentProjectInteractionFactory(InteractionFactoryBase):
 
     kind = Interaction.KINDS.interaction
     investment_project = factory.SubFactory(InvestmentProjectFactory)
-    communication_channel_id = constants.CommunicationChannel.face_to_face.value.id
+    communication_channel = factory.LazyFunction(
+        lambda: random_obj_for_model(CommunicationChannel)
+    )
 
 
 class ServiceDeliveryFactory(InteractionFactoryBase):
