+++ datahub/search/interaction/tests/test_views.py
@@ -12,6 +12,7 @@ from rest_framework.reverse import reverse
 from datahub.company.test.factories import AdviserFactory, CompanyFactory, ContactFactory
 from datahub.core import constants
 from datahub.core.test_utils import APITestMixin, create_test_user
+from datahub.interaction.constants import CommunicationChannel
 from datahub.interaction.models import Interaction
 from datahub.interaction.test.factories import CompanyInteractionFactory, ServiceDeliveryFactory
 from datahub.metadata.test.factories import TeamFactory
@@ -471,9 +472,9 @@ class TestViews(APITestMixin):
         """Tests filtering interaction by interaction type."""
         CompanyInteractionFactory.create_batch(
             5,
-            communication_channel_id=constants.CommunicationChannel.email_website.value.id
+            communication_channel_id=CommunicationChannel.email_website.value.id
         )
-        communication_channel_id = constants.CommunicationChannel.social_media.value.id
+        communication_channel_id = CommunicationChannel.social_media.value.id
         CompanyInteractionFactory.create_batch(
             5,
             communication_channel_id=communication_channel_id
